Moved all clockreduction tests and cleaned them up. Moved creation of ClockAnalysisGraph into constructor. Added a simpler way to access CompiledComponents declarations.
