--- src/tests/clock_reduction/mod.rs
@@ -1,6 +1,0 @@
-mod advanced_clock_detection_test;
-mod advanced_clock_removal_test;
-mod clock_removal_test;
-mod helper;
-pub mod redundant_clock_detection_test;
-pub mod unused_clock_detection_test;
